+++ summoner.mjs
@@ -13,7 +13,7 @@ Hooks.once("init", function() {
 
 Hooks.once("setup", function() {
   libWrapper.register(
-    "arbron-summoner", "Item.implementation.prototype.getRollData", SummonsItem.getRollData, libWrapper.WRAPPER
+    "arbron-summoner", "Actor.implementation.prototype.getRollData", SummonsActor.getRollData, libWrapper.WRAPPER
   );
 });